src/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
azure-core==1.26.1
azure-storage-blob==12.14.1
black==22.10.0
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coverage==6.5.0
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.51.0
django-storages==1.13.1
exceptiongroup==1.0.4
hiredis==2.0.0
idna==3.4
iniconfig==1.1.1
isodate==0.6.1
mccabe==0.7.0
msrest==0.7.1
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
python3-openid==3.2.0
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
whitenoise==6.2.0
